Extract ClientMsg and ServerMsg
