+++ packages/liveblocks-core/src/__tests__/room.test.ts
@@ -12,12 +12,14 @@ import {
   createRoom,
 } from "../room";
 import type { Authentication, Others } from "../types";
-import { ClientMsgCode, ServerMsgCode, WebsocketCloseCodes } from "../types";
+import { WebsocketCloseCodes } from "../types";
 import type { BaseUserMeta } from "../types/BaseUserMeta";
+import { ClientMsgCode } from "../types/ClientMsg";
 import type { Json, JsonObject } from "../types/Json";
 import type { LsonObject } from "../types/Lson";
 import type { IdTuple, SerializedCrdt } from "../types/SerializedCrdt";
 import { CrdtType } from "../types/SerializedCrdt";
+import { ServerMsgCode } from "../types/ServerMsg";
 import {
   createSerializedList,
   createSerializedObject,
