+++ packages/liveblocks-core/src/types/index.ts
@@ -741,25 +741,6 @@ export enum WebsocketCloseCodes {
   CLOSE_WITHOUT_RETRY = 4999,
 }
 
-export type {
-  BroadcastEventClientMsg,
-  ClientMsg,
-  FetchStorageClientMsg,
-  UpdatePresenceClientMsg,
-  UpdateStorageClientMsg,
-} from "./ClientMsg";
-export { ClientMsgCode } from "./ClientMsg";
 export type { Immutable, ToImmutable } from "./Immutable";
 export type { Json, JsonObject } from "./Json";
 export type { LiveNode, LiveStructure, Lson, LsonObject, ToJson } from "./Lson";
-export type {
-  BroadcastedEventServerMsg,
-  InitialDocumentStateServerMsg,
-  RoomStateServerMsg,
-  ServerMsg,
-  UpdatePresenceServerMsg,
-  UpdateStorageServerMsg,
-  UserJoinServerMsg,
-  UserLeftServerMsg,
-} from "./ServerMsg";
-export { ServerMsgCode } from "./ServerMsg";
